corrected the fdlf, and structured the project
