+++ Decoupled/fdlf.py
@@ -1,0 +1,10 @@
+import sys
+sys.path.append(".")
+from Decoupled.primal import *
+from Decoupled.dual import *
+
+
+def fdlf_print(Pactual,Qactual,Pindex, Qindex, tindex, vindex, v, teta, g, b,z):
+    if(print_primal(Pactual,Qactual,Pindex, Qindex, tindex, vindex, v, teta, g, b,z)==0):
+        print_dual(Pactual,Qactual,Pindex, Qindex, tindex, vindex, v, teta, g, b,z)
+
