+++ Decoupled/optionally.py
@@ -1,29 +1,31 @@
-import Loadflow.PowerEq as pe
+import sys
+sys.path.append(".")
+import Loadflow.voltstab as lf
 import numpy as np
 
 
 def h_matrix(Pindex, tindex, v, teta, g, b):
     matrix=np.zeros(shape=(Pindex.size,tindex.size))
     for i in Pindex:
-            matrix[i-1]=pe.dpdt(i,tindex,v,teta,g,b)
+            matrix[i-1]=lf.dpdt(i,tindex,v,teta,g,b)
     return matrix
 
 def n_matrix(Pindex, vindex, v, teta, g, b):
     matrix = np.zeros(shape=(Pindex.size, vindex.size))
     for i in Pindex:
-        matrix[i - 1] = pe.dpdv(i, vindex, v, teta, g, b)
+        matrix[i - 1] = lf.dpdv(i, vindex, v, teta, g, b)
     return matrix
 
 def m_matrix(Qindex, tindex, v, teta, g, b):
     matrix = np.zeros(shape=(Qindex.size, tindex.size))
     for i in Qindex:
-        matrix[i - 1] = pe.dqdt(i, tindex, v, teta, g, b)
+        matrix[i - 1] = lf.dqdt(i, tindex, v, teta, g, b)
     return matrix
 
 def l_matrix(Qindex, vindex, v, teta, g, b):
     matrix = np.zeros(shape=(Qindex.size, vindex.size))
     for i in Qindex:
-        matrix[i - 1] = pe.dqdv(i, vindex, v, teta, g, b)
+        matrix[i - 1] = lf.dqdv(i, vindex, v, teta, g, b)
     return matrix
 
 def heq_matrix(Pindex,Qindex,tindex,vindex,v,teta,g,b):
